requirements.txt
opencv-python
onvif-zeep
pyyaml
